Add scripts to save data directly through pandas
This method was chosen to avoid having to create large fixture files, which
was space-reduntant and very slow. Now, the pandas to_sql command directly
stores data in MySQL tables, which is much faster

The package django-environ was replaced with python-dotenv because the
.env file is now also being used in the store_data.py script. python-dotenv
does a better job at finding the .env file in the root of the project.

Finally, a requirements.txt file was created and will now always be updated