Build basic test cases and add defaul model values
The purpose of these test cases was to verify that the many-to-many
relationship between keywords and movies is working.

The documentation was also updated with instructions on creating a test
database to run the test suite in.

Finally, tagline, title, and vote_average were given default values to
accomodate for incomplete movie data. This change was inspired by the
fact that test objects could not be created without populating the
vote_average field.