Add ability to add search match scores
These match scores are currently stored in-memory within the SummaryInformation
class. This may be changed in the future due to a sorting problem. In order
to sort movies by their match score, the score must be a field in the movie
model. In-memory storage of match scores does not allow for sorting.