Add images to movie pages on React app
This update uses the Google Custom Search Engine. An api call to
GCSE is made in the Django backend. The Django backend exposes this
call as its own endpoint. The React frontend calls Django's image
endpoint and displays the first image on the movie page.

GCSE API calls cannot be called by a client server. This is why, the
call is made through Django, which is then called by React